Make shared.rs to eliminate some cycles in the dependency graph.
